refactor: rename getCardAriaLabel to getReadMoreAriaLabel and move to dedicated file
