+++ src/components/ItaliaTheme/Blocks/ContactsBlock/View.jsx
@@ -12,7 +12,7 @@ import ViewBlock from './Block/ViewBlock';
 import { Container, Row, Col } from 'design-react-kit/dist/design-react-kit';
 import { flattenToAppURL } from '@plone/volto/helpers';
 import { UniversalLink } from '@plone/volto/components';
-import { getCardAriaLabel } from 'design-comuni-plone-theme/helpers';
+import { getReadMoreAriaLabel } from 'design-comuni-plone-theme/helpers';
 import config from '@plone/volto/registry';
 
 /**
@@ -26,7 +26,7 @@ const AccordionView = ({ data, block }) => {
   const blockPlainTitle = data.title
     ? convertFromRaw(data.title).getPlainText()
     : null;
-  const linkMoreAriaLabel = getCardAriaLabel(intl, blockPlainTitle);
+  const linkMoreAriaLabel = getReadMoreAriaLabel(intl, blockPlainTitle);
   return (
     <div className="block contacts">
       <div className="public-ui">
