+++ src/components/ItaliaTheme/Blocks/HeroImageLeft/StoresButtons.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { Button } from 'design-react-kit/dist/design-react-kit';
 import { defineMessages, useIntl } from 'react-intl';
-import { getCardAriaLabel } from 'design-comuni-plone-theme/helpers';
+import { getReadMoreAriaLabel } from 'design-comuni-plone-theme/helpers';
 import { Icon } from 'design-comuni-plone-theme/components/ItaliaTheme';
 import { UniversalLink } from '@plone/volto/components';
 import { CardReadMore } from 'design-react-kit/dist/design-react-kit';
@@ -41,7 +41,7 @@ const StoresButtons = ({ data }) => {
           iconName="it-arrow-right"
           text={data.moreTitle || intl.formatMessage(messages.linkMore)}
           href={flattenToAppURL(data.moreHref)}
-          aria-label={getCardAriaLabel(intl, data.title)}
+          aria-label={getReadMoreAriaLabel(intl, data.title)}
         />
       )}
     </div>
