+++ src/components/ItaliaTheme/Blocks/Listing/BandiInEvidenceTemplate.jsx
@@ -17,7 +17,10 @@ import {
   ListingText,
   ListingLinkMore,
 } from 'design-comuni-plone-theme/components/ItaliaTheme';
-import { viewDate, getCardAriaLabel } from 'design-comuni-plone-theme/helpers';
+import {
+  viewDate,
+  getReadMoreAriaLabel,
+} from 'design-comuni-plone-theme/helpers';
 
 const messages = defineMessages({
   vedi: {
@@ -251,7 +254,7 @@ const BandiInEvidenceTemplate = ({
                       item={!isEditMode ? item : null}
                       href={isEditMode ? '#' : null}
                       text={intl.formatMessage(messages.vedi)}
-                      aria-label={getCardAriaLabel(intl, item.title)}
+                      aria-label={getReadMoreAriaLabel(intl, item.title)}
                     />
                   </div>
                 </CardBody>
