+++ src/helpers/index.js
@@ -11,10 +11,8 @@ export {
   getEventRecurrenceMore,
 } from 'design-comuni-plone-theme/helpers/ListingHelper';
 export { contentFolderHasItems } from 'design-comuni-plone-theme/helpers/contentHelper';
-export {
-  checkRedraftHasContent,
-  getCardAriaLabel,
-} from 'design-comuni-plone-theme/helpers/redraftHelper';
+export { checkRedraftHasContent } from 'design-comuni-plone-theme/helpers/redraftHelper';
+export { getReadMoreAriaLabel } from 'design-comuni-plone-theme/helpers/getReadMoreAriaLabel';
 export { getTableRowData } from 'design-comuni-plone-theme/helpers/amministrazioneTrasparenteHelper';
 export { getItemsByPath } from 'design-comuni-plone-theme/helpers/getItemsByPath';
 export {
